move basic pf types to versions.types
