+++ powerfactory_tools/versions/pf2022/types.py
@@ -12,98 +12,8 @@ if t.TYPE_CHECKING:
     from collections.abc import Sequence
 
     from powerfactory_tools.versions.types import Currency
+    from powerfactory_tools.versions.types import FolderType
     from powerfactory_tools.versions.types import MetricPrefix
-
-
-class PFClassId(enum.Enum):
-    AREA = "ElmArea"
-    COMPOUND_GRID_ELEMENT = "ElmFolder"  # e.g. a compound grid graphic consisting of multiple elements
-    COMPOUND_MODEL = "ElmComp"  # e.g. a compound generator with multiple functional slots as part of a template model
-    COUPLER = "ElmCoup"
-    CUBICLE = "StaCubic"
-    CURRENT_SOURCE_AC = "ElmIac"
-    EXTERNAL_GRID = "ElmXNet"
-    DATETIME = "SetTime"
-    DESKTOP = "SetDesktop"
-    DSL_MODEL = "ElmDsl"
-    FOLDER = "IntFolder"
-    FUSE = "RelFuse"
-    FUSE_TYPE = "TypFuse"
-    GENERATOR = "ElmGenstat"
-    GRID = "ElmNet"
-    GRID_GRAPHIC = "IntGrfnet"
-    LINE = "ElmLne"
-    LINE_TYPE = "TypLne"
-    LOAD = "ElmLod"
-    LOAD_LV = "ElmLodlv"
-    LOAD_LV_PART = "ElmLodlvp"
-    LOAD_MV = "ElmLodmv"
-    LOAD_TYPE_GENERAL = "TypLod"
-    LOAD_TYPE_LV = "TypLodlv"
-    LOAD_TYPE_MV = "TypDistrf"
-    MEASUREMENT_FILE = "ElmFile"
-    PROJECT_FOLDER = "IntPrjfolder"
-    PROJECT_SETTINGS = "SetPrj"
-    PVSYSTEM = "ElmPvsys"
-    REFERENCE = "IntRef"
-    RESULT = "ElmRes"
-    SCENARIO = "IntScenario"
-    SCRIPT_DPL = "ComDpl"
-    SCRIPT_PYTHON = "ComPython"
-    SELECTION = "SetSelect"
-    SETTINGS_FOLDER = "SetFold"
-    SETTINGS_FOLDER_UNITS = "IntUnit"
-    SOURCE_TYPE_HARMONIC_CURRENT = "TypHmccur"
-    STATION_CONTROLLER = "ElmStactrl"
-    STUDY_CASE = "IntCase"
-    SWITCH = "StaSwitch"
-    TEMPLATE = "IntTemplate"
-    TERMINAL = "ElmTerm"
-    TRANSFORMER_2W = "ElmTr2"
-    TRANSFORMER_2W_TYPE = "TypTr2"
-    TRANSFORMER_3W = "ElmTr3"
-    TRANSFORMER_3W_TYPE = "TypTr3"
-    UNIT_VARIABLE = "SetVariable"
-    VARIABLE_MONITOR = "IntMon"  # Variable monitor definition
-    VARIANT = "IntScheme"
-    VARIANT_CONFIG = "IntAcscheme"
-    VARIANT_STAGE = "IntSstage"
-    ZONE = "ElmZone"
-
-
-class FolderType(enum.Enum):
-    CB_RATINGS = "cbrat"
-    CIM_MODEL = "cim"
-    CHARACTERISTICS = "chars"
-    COMMON_MODE_FAILURES = "common"
-    DEMAND_TRANSFERS = "demand"
-    DIAGRAMS = "dia"
-    EQUIPMENT_TYPE_LIBRARY = "equip"
-    FAULTS = "fault"
-    GENERIC = "gen"
-    GENERATOR_COST_CURVES = "cstgen"
-    GENERATOR_EFFICIENCY_CURVES = "effgen"
-    LIBRARY = "lib"
-    MVAR_LIMIT_CURVES = "mvar"
-    NETWORK_DATA = "netdat"
-    NETWORK_MODEL = "netmod"
-    OPERATIONAL_LIBRARY = "oplib"
-    OPERATION_SCENARIOS = "scen"
-    OUTAGES = "outage"
-    QP_CURVES = "qpc"
-    PROBABILISTIC_ASSESSMENT = "rnd"
-    RUNNING_ARRANGEMENTS = "ra"
-    REMEDIAL_ACTION_SCHEMES = "ras"
-    SCRIPTS = "script"
-    STATION_WARE = "sw"
-    STUDY_CASES = "study"
-    TABLE_REPORTS = "report"
-    TARIFFS = "tariff"
-    TEMPLATES = "templ"
-    THERMAL_RATINGS = "therm"
-    USER_DEFINED_MODELS = "blk"
-    VARIATIONS = "scheme"
-    V_CONTROL_CURVES = "ucc"
 
 
 class LocalQCtrlMode(enum.Enum):
